+++ AAA_WEBAPP/assets/js/values.js
@@ -1,4 +1,4 @@
 const VALUES = {
 	assets: "/assets/",
-	api: "http://spotify-labelling-api.21webb.nl/",
+	api: "http://spotify-labelling-api.21webb.nl/"
 };