+++ AAA_WEBAPP/assets/snippets/css.php
@@ -1,0 +1,7 @@
+<link rel="stylesheet" type="text/css" href="/assets/css/general.css">
+<link rel="stylesheet" type="text/css" href="/assets/css/dashboard.css">
+<link rel="stylesheet" type="text/css" href="/assets/css/form.css">
+<link rel="stylesheet" type="text/css" href="/assets/css/lazy-loading.css">
+<link rel="stylesheet" type="text/css" href="/assets/css/module.css">
+<link rel="stylesheet" type="text/css" href="/assets/css/navigation.css">
+<link rel="stylesheet" type="text/css" href="/assets/css/popup.css">