+++ AAA_WEBAPP/dashboard/songs/index.php
@@ -13,8 +13,7 @@ session_start();
 		
 		<link rel="preconnect" href="https://fonts.gstatic.com">
 		<link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Fjalla+One&display=swap">
-		<link rel="stylesheet" type="text/css" href="../../assets/css/general.css">
-		<link rel="stylesheet" type="text/css" href="../../assets/css/dashboard.css">
+		<?php include_once("../../assets/snippets/css.php"); ?>
 	</head>
 
 
@@ -35,15 +34,6 @@ session_start();
 			<!-- Content -->
 			<div class="module">
 
-				<button class="wide" id="import-songs">IMPORT SONGS</button>
-
-				<div class="table-container">
-					<table>
-						<tbody id="songs">
-						</tbody>
-					</table>
-				</div>
-
 			</div>
 
 		</div>	<!-- .main-wrapper -->
@@ -59,49 +49,6 @@ session_start();
 
 		// Protect the page
 		PageProtect.protect({ verifiedLevel: 2 });
-
-		// Add "Import songs" button functionality
-		document.getElementById("import-songs").addEventListener("click", async () => {
-
-
-
-			// ============================================================================
-			// TODO: THINK OF A SMARTER WAY TO DO THIS
-			// A TABLE DOESN'T WORK WELL WITH HTML JS FORMS
-			// SO THE SOLUTION MUST WORK WITH THE CURRENT JAVASCRIPT FILES
-			// WELL ENOUGH TO ONLY WRITE ADDITIONAL CODE INSTEAD OF REWRITING PARTS
-			// BUT IT ALSO NEEDS EITHER TABLE SUPPORT OR ANOTHER WAY
-			// OF DISPLAYING THE DATA TO THE USER
-			// ADDITIONAL PAGES IS NOT TOO BAD IF THAT IS NEEDED
-			// ============================================================================
-
-
-
-
-			// Create popup
-			const importSongs = new BigPopup("Import Songs", "api/v1/spotify/import", "POST", "import-songs-form");
-
-			const tableContainer = Api.createElement("div", {classList: "table-container limit-height"});
-			const table = Api.createElement("table");
-			
-			// Show all playlists for the user to choose from
-			const res = await Api.sendRequest("api/v1/spotify/playlists", "GET");
-			for(const list of res.data.playlists) {
-				
-				const row = Api.createElement("tr");
-				row.appendChild(Api.createElement("td", { innerHTML: list.name }));
-				row.appendChild(Api.createElement("td", { innerHTML: list.numTracks + " songs" }));
-				row.appendChild(Api.createElement("input", { type: "checkbox", name: "input id-" + list.spotifyID }));
-				table.appendChild(row);
-
-			}
-
-			tableContainer.appendChild(table);
-			importSongs.addElement(tableContainer);
-			importSongs.show("IMPORT");
-			HtmlJsForm.findById("import-songs-form").addCallback(() => {  });
-
-		});
 
 		</script>