+++ AAA_WEBAPP/register/index.php
@@ -10,7 +10,7 @@
 
 		<link rel="preconnect" href="https://fonts.gstatic.com">
 		<link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Fjalla+One&display=swap">
-		<link rel="stylesheet" type="text/css" href="../assets/css/general.css">
+		<?php include_once("../assets/snippets/css.php"); ?>
 
 	</head>